More standard API for function calls
Made most CPrAN functions object methods. This meant moving
some of those "global" functions to the specific commands
that really needed them (e.g., known() and installed() were
moved to "search"), and modifying the calls to those functions
so that they either called the commands, called their own
methods instead of these globals, or called the globals only
when the values could not be obtained from the provided
options. This is necessary when custom paths are passed.