Try to grab version string from binary
The previous method for detecting Praat versions worked well,
but it required reading from STDOUT. When implementing support
for piping arguments from other commands, this proved to be
insufficient.

This method works well in existing versions of Praat for both
GNU/Linux and Windows, but it hasn't yet been tested in Mac.